[`pylint`] Implement `too-many-try-statements` (`W0717`)
