+++ crates/ruff_linter/src/rules/pylint/helpers.rs
@@ -1,4 +1,5 @@
 use ruff_python_ast as ast;
+use ruff_python_ast::ExceptHandler;
 use ruff_python_ast::visitor::Visitor;
 use ruff_python_ast::{Arguments, Expr, Stmt, visitor};
 use ruff_python_semantic::analyze::function_type;
@@ -314,4 +315,387 @@ pub(super) fn is_known_dunder_method(method: &str) -> bool {
             | "_order_"
             | "_generate_next_value_"
         )
+}
+
+pub(super) fn num_statements(stmts: &[Stmt]) -> usize {
+    let mut count = 0;
+    for stmt in stmts {
+        match stmt {
+            Stmt::If(ast::StmtIf {
+                body,
+                elif_else_clauses,
+                ..
+            }) => {
+                count += 1;
+                count += num_statements(body);
+                for clause in elif_else_clauses {
+                    count += 1;
+                    count += num_statements(&clause.body);
+                }
+            }
+            Stmt::For(ast::StmtFor { body, orelse, .. }) => {
+                count += num_statements(body);
+                count += num_statements(orelse);
+            }
+            Stmt::While(ast::StmtWhile { body, orelse, .. }) => {
+                count += 1;
+                count += num_statements(body);
+                count += num_statements(orelse);
+            }
+            Stmt::Match(ast::StmtMatch { cases, .. }) => {
+                count += 1;
+                for case in cases {
+                    count += 1;
+                    count += num_statements(&case.body);
+                }
+            }
+            Stmt::Try(ast::StmtTry {
+                body,
+                handlers,
+                orelse,
+                finalbody,
+                ..
+            }) => {
+                count += 1;
+                count += num_statements(body);
+                if !orelse.is_empty() {
+                    count += 1 + num_statements(orelse);
+                }
+                if !finalbody.is_empty() {
+                    // Unclear why, but follow Pylint's convention.
+                    count += 2 + num_statements(finalbody);
+                }
+                if handlers.len() > 1 {
+                    count += 1;
+                }
+                for handler in handlers {
+                    count += 1;
+                    let ExceptHandler::ExceptHandler(ast::ExceptHandlerExceptHandler {
+                        body, ..
+                    }) = handler;
+                    count += num_statements(body);
+                }
+            }
+            Stmt::FunctionDef(ast::StmtFunctionDef { body, .. })
+            | Stmt::With(ast::StmtWith { body, .. }) => {
+                count += 1;
+                count += num_statements(body);
+            }
+            Stmt::Return(_) => {}
+            _ => {
+                count += 1;
+            }
+        }
+    }
+    count
+}
+
+#[cfg(test)]
+mod tests {
+    use anyhow::Result;
+
+    use ruff_python_ast::Suite;
+    use ruff_python_parser::parse_module;
+
+    use super::num_statements;
+
+    fn parse_suite(source: &str) -> Result<Suite> {
+        Ok(parse_module(source)?.into_suite())
+    }
+
+    #[test]
+    fn pass() -> Result<()> {
+        let source: &str = r"
+def f():
+    pass
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 2);
+        Ok(())
+    }
+
+    #[test]
+    fn if_else() -> Result<()> {
+        let source: &str = r"
+def f():
+    if a:
+        print()
+    else:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 5);
+        Ok(())
+    }
+
+    #[test]
+    fn if_else_if_corner() -> Result<()> {
+        let source: &str = r"
+def f():
+    if a:
+        print()
+    else:
+        if a:
+            print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 6);
+        Ok(())
+    }
+
+    #[test]
+    fn if_elif() -> Result<()> {
+        let source: &str = r"
+def f():
+    if a:
+        print()
+    elif a:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 5);
+        Ok(())
+    }
+
+    #[test]
+    fn if_elif_else() -> Result<()> {
+        let source: &str = r"
+def f():
+    if a:
+        print()
+    elif a == 2:
+        print()
+    elif a == 3:
+        print()
+    else:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 9);
+        Ok(())
+    }
+
+    #[test]
+    fn match_case() -> Result<()> {
+        let source: &str = r"
+def f():
+    match x:
+        case 3:
+            pass
+        case _:
+            pass
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 6);
+        Ok(())
+    }
+
+    #[test]
+    fn many_statements() -> Result<()> {
+        let source: &str = r"
+async def f():
+    a = 1
+    b = 2
+    c = 3
+    await some_other_func()
+    if a == 1:
+        print('hello')
+    else:
+        other_func()
+    count = 1
+    while True:
+        count += 1
+        if count > 20:
+            break;
+
+    with open(f):
+        with open(e):
+            a -= 1
+            import time
+            pass
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 19);
+        Ok(())
+    }
+
+    #[test]
+    fn for_() -> Result<()> {
+        let source: &str = r"
+def f():
+    for i in range(10):
+        pass
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 2);
+        Ok(())
+    }
+
+    #[test]
+    fn for_else() -> Result<()> {
+        let source: &str = r"
+def f():
+    for i in range(10):
+        print()
+    else:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 3);
+        Ok(())
+    }
+
+    #[test]
+    fn nested_def() -> Result<()> {
+        let source: &str = r"
+def f():
+    def g():
+        print()
+        print()
+
+    print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 5);
+        Ok(())
+    }
+
+    #[test]
+    fn nested_class() -> Result<()> {
+        let source: &str = r"
+def f():
+    class A:
+        def __init__(self):
+            pass
+
+        def f(self):
+            pass
+
+    print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 3);
+        Ok(())
+    }
+
+    #[test]
+    fn return_not_counted() -> Result<()> {
+        let source: &str = r"
+def f():
+    return
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 1);
+        Ok(())
+    }
+
+    #[test]
+    fn with() -> Result<()> {
+        let source: &str = r"
+def f():
+    with a:
+        if a:
+            print()
+        else:
+            print()
+
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 6);
+        Ok(())
+    }
+
+    #[test]
+    fn try_except() -> Result<()> {
+        let source: &str = r"
+def f():
+    try:
+        print()
+    except Exception:
+        raise
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 5);
+        Ok(())
+    }
+
+    #[test]
+    fn try_except_else() -> Result<()> {
+        let source: &str = r"
+def f():
+    try:
+        print()
+    except ValueError:
+        pass
+    else:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 7);
+        Ok(())
+    }
+
+    #[test]
+    fn try_except_else_finally() -> Result<()> {
+        let source: &str = r"
+def f():
+    try:
+        print()
+    except ValueError:
+        pass
+    else:
+        print()
+    finally:
+        pass
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 10);
+        Ok(())
+    }
+
+    #[test]
+    fn try_except_except() -> Result<()> {
+        let source: &str = r"
+def f():
+    try:
+        print()
+    except ValueError:
+        pass
+    except Exception:
+        raise
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 8);
+        Ok(())
+    }
+
+    #[test]
+    fn try_except_except_finally() -> Result<()> {
+        let source: &str = r"
+def f():
+    try:
+        print()
+    except:
+        pass
+    except:
+        pass
+    finally:
+        print()
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 11);
+        Ok(())
+    }
+
+    #[test]
+    fn yield_() -> Result<()> {
+        let source: &str = r"
+def f():
+    for i in range(10):
+        yield i
+";
+        let stmts = parse_suite(source)?;
+        assert_eq!(num_statements(&stmts), 2);
+        Ok(())
+    }
 }
