+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -90,6 +90,7 @@ pub(crate) use too_many_positional_arguments::*;
 pub(crate) use too_many_public_methods::*;
 pub(crate) use too_many_return_statements::*;
 pub(crate) use too_many_statements::*;
+pub(crate) use too_many_try_statements::*;
 pub(crate) use type_bivariance::*;
 pub(crate) use type_name_incorrect_variance::*;
 pub(crate) use type_param_name_mismatch::*;
@@ -202,6 +203,7 @@ mod too_many_positional_arguments;
 mod too_many_public_methods;
 mod too_many_return_statements;
 mod too_many_statements;
+mod too_many_try_statements;
 mod type_bivariance;
 mod type_name_incorrect_variance;
 mod type_param_name_mismatch;
